A bug that prevents the ban account details and phone number to be entered to the database
Turns out that the value was exceeding the amount for the integer data type.
All have been corrected to  long data type and database has been updated to BIGINT as oppsed to INTEGER from before

+++ src/database/PurchaseConfirmedSQL.java
@@ -21,8 +21,8 @@ public class PurchaseConfirmedSQL {
                                String lastName,
                                String permanentAddress,
                                String currentAddress,
-                               int contactNumber,
-                               int bankAccountNum) throws ClassNotFoundException, SQLException{
+                               long contactNumber,
+                               long bankAccountNum) throws ClassNotFoundException, SQLException{
         
         if(checkCustomerExists(identityNumber) == 0){ //check if the customer exists
             String insertToCustomer = "INSERT INTO customers "
